_code-samples/airgapped-wallet/py/requirements.txt
anyio==3.2.1
asgiref==3.4.1
base58==2.1.0
certifi==2023.7.22
cffi==1.15.0
colorama==0.4.4
cryptography==41.0.6
Django==3.2.23
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.6
idna==3.2
image==1.5.33
pifacedigitalio==3.0.5
Pillow==10.2.0
pycparser==2.20
pytz==2021.1
qrcode==7.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.4
typing-extensions==4.2.0
websockets==10.0
xrpl-py==2.0.0
